add helpdesk notification email on new ticket & update auth

+++ app/Http/Services/Auth/AuthCrudService.php
@@ -31,6 +31,8 @@ class AuthCrudService
      */
     public function logoutUser($token): void
     {
-        $token->delete();
+        if ($token && method_exists($token, 'delete')) {
+            $token->delete();
+        }
     }
 }
